# Frontend/sssklv2-openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "SSSKLv2 | v1",
        "contact": {
            "name": "Scouting Wilo",
            "email": "webmaster@scoutingwilo.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:7193/"
        }
    ],
    "paths": {
...
        "/v1/Events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 15
                        }
                    },
                    {
                        "name": "futureOnly",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginationObjectOfEventDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginationObjectOfEventDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginationObjectOfEventDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "type": "object",
                                        "properties": {
                                            "Title": {
                                                "type": "string"
                                            },
                                            "Description": {
                                                "type": "string"
                                            },
                                            "ImageContent.CanRead": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanWrite": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanSeek": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanTimeout": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.Length": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "ImageContent.Position": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "ImageContent.ReadTimeout": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int32"
                                            },
                                            "ImageContent.WriteTimeout": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int32"
                                            },
                                            "ImageContentType.Boundary": {
                                                "type": "string"
                                            },
                                            "ImageContentType.CharSet": {
                                                "type": "string"
                                            },
                                            "ImageContentType.MediaType": {
                                                "type": "string"
                                            },
                                            "ImageContentType.Name": {
                                                "type": "string"
                                            },
                                            "ImageContentType.Parameters": {
                                                "type": "array"
                                            },
                                            "StartDateTime": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "EndDateTime": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "ContentType": {
                                                "type": "string"
                                            },
                                            "ContentDisposition": {
                                                "type": "string"
                                            },
                                            "Headers": {
                                                "type": "object",
                                                "additionalProperties": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "Length": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "Name": {
                                                "type": "string"
                                            },
                                            "FileName": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/Events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "type": "object",
                                        "properties": {
                                            "Title": {
                                                "type": "string"
                                            },
                                            "Description": {
                                                "type": "string"
                                            },
                                            "ImageContent.CanRead": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanWrite": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanSeek": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.CanTimeout": {
                                                "type": "boolean"
                                            },
                                            "ImageContent.Length": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "ImageContent.Position": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "ImageContent.ReadTimeout": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int32"
                                            },
                                            "ImageContent.WriteTimeout": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int32"
                                            },
                                            "ImageContentType.Boundary": {
                                                "type": "string"
                                            },
                                            "ImageContentType.CharSet": {
                                                "type": "string"
                                            },
                                            "ImageContentType.MediaType": {
                                                "type": "string"
                                            },
                                            "ImageContentType.Name": {
                                                "type": "string"
                                            },
                                            "ImageContentType.Parameters": {
                                                "type": "array"
                                            },
                                            "StartDateTime": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "EndDateTime": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "ContentType": {
                                                "type": "string"
                                            },
                                            "ContentDisposition": {
                                                "type": "string"
                                            },
                                            "Headers": {
                                                "type": "object",
                                                "additionalProperties": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "Length": {
                                                "pattern": "^-?(?:0|[1-9]\\d*)$",
                                                "type": [
                                                    "integer",
                                                    "string"
                                                ],
                                                "format": "int64"
                                            },
                                            "Name": {
                                                "type": "string"
                                            },
                                            "FileName": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/Events/{id}/rsvp": {
            "post": {
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventResponseDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventResponseDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventResponseDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EventDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "startDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "creatorName": {
                        "type": "string"
                    },
                    "createdOn": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "acceptedUsers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EventResponseUserDto"
                        }
                    },
                    "declinedUsers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EventResponseUserDto"
                        }
                    },
                    "userResponse": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/EventResponseStatus"
                            }
                        ]
                    }
                }
            },
            "EventResponseDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/EventResponseStatus"
                    }
                }
            },
            "EventResponseStatus": {
                "enum": [
                    "Accepted",
                    "Declined",
                    null
                ]
            },
            "EventResponseUserDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    }
                }
            },
...
            "PaginationObjectOfEventDto": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EventDto"
                        }
                    },
                    "totalCount": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    }
                }
            },
...
        }
    },
...
}